pagesnapcontainer component
