+++ src/components/fullpage.js
@@ -7,8 +7,8 @@ const FullPage = ({ children }) => (
   <div className={styles.fullPage}>{children}</div>
 )
 
-export default FullPage
-
 FullPage.propTypes = {
   children: PropTypes.node.isRequired,
 }
+
+export default FullPage
