+++ src/components/pagesnap.module.styl
@@ -1,0 +1,2 @@
+.page-snap-container
+  height: 100vh
